+++ connectors/src/lib/models/microsoft.ts
@@ -6,28 +6,9 @@ import type {
 } from "sequelize";
 import { DataTypes, Model } from "sequelize";
 
+import type { MicrosoftNodeType } from "@connectors/connectors/microsoft/lib/node_types";
 import { sequelizeConnection } from "@connectors/resources/storage";
 import { ConnectorModel } from "@connectors/resources/storage/models/connector_model";
-
-export const MICROSOFT_NODE_TYPES = [
-  "sites-root",
-  "teams-root",
-  "site",
-  "team",
-  "drive",
-  "folder",
-  "file",
-  "page",
-  "channel",
-  "message",
-] as const;
-export type MicrosoftNodeType = (typeof MICROSOFT_NODE_TYPES)[number];
-
-export function isValidNodeType(
-  nodeType: string
-): nodeType is MicrosoftNodeType {
-  return MICROSOFT_NODE_TYPES.includes(nodeType as MicrosoftNodeType);
-}
 
 export class MicrosoftConfigurationModel extends Model<
   InferAttributes<MicrosoftConfigurationModel>,
